# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Emumet Account Service API",
    "description": "Account Service for ShuttlePub",
    "contact": {
      "name": "turtton"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/accounts": {
      "get": {
        "tags": [
          "Account"
        ],
        "description": "Retrieve accounts by IDs or with cursor-based pagination.",
        "operationId": "get_accounts",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "Comma-separated account IDs",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination limit",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "description": "Pagination direction (asc/desc)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "404": {
            "description": "No accounts found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Account"
        ],
        "description": "Create a new account with a generated signing key pair.",
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/accounts/{id}": {
      "put": {
        "tags": [
          "Account"
        ],
        "description": "Update account properties.",
        "operationId": "update_account_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Account updated"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Account"
        ],
        "description": "Deactivate an account and cascade-delete related resources.",
        "operationId": "deactivate_account_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Account deactivated"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/accounts/{id}/ban": {
      "post": {
        "tags": [
          "Account"
        ],
        "description": "Permanently ban an account.",
        "operationId": "ban_account_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Account banned"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/accounts/{id}/suspend": {
      "post": {
        "tags": [
          "Account"
        ],
        "description": "Suspend an account with a reason and optional expiry.",
        "operationId": "suspend_account_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Account suspended"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/accounts/{id}/unsuspend": {
      "post": {
        "tags": [
          "Account"
        ],
        "description": "Remove suspension from an account.",
        "operationId": "unsuspend_account_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Account unsuspended"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "public_key",
          "is_bot",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "is_bot": {
            "type": "boolean"
          },
          "moderation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModerationResponse"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          }
        }
      },
      "AccountsResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "first": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            }
          },
          "last": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BanAccountRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "CreateAccountRequest": {
        "type": "object",
        "required": [
          "name",
          "is_bot"
        ],
        "properties": {
          "is_bot": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ModerationResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "reason",
              "suspended_at",
              "type"
            ],
            "properties": {
              "expires_at": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              },
              "reason": {
                "type": "string"
              },
              "suspended_at": {
                "type": "string",
                "format": "date-time"
              },
              "type": {
                "type": "string",
                "enum": [
                  "suspended"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reason",
              "banned_at",
              "type"
            ],
            "properties": {
              "banned_at": {
                "type": "string",
                "format": "date-time"
              },
              "reason": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "banned"
                ]
              }
            }
          }
        ]
      },
...
      "SuspendAccountRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "UpdateAccountRequest": {
        "type": "object",
        "required": [
          "is_bot"
        ],
        "properties": {
          "is_bot": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}